requirements.txt
pycryptodome==3.9.7
aioconsole==0.1.16